Setup project
I'm just setting up the code for the project, and putting some of the
preferred tooling around the project. I still need to write a README and
configure the rest of uploading to PyPI. I also want to re-visit the
code itself and see if the quality can be improved, and to re-visit the
tests to see if there's stuff there that can be improved as well.