Support recursive ForwardRef type declarations
In Python, if you want to use the class that you're defining, within the
class itself, you have to wrap the class in quotes so that you can do
what is known as a ForwardRef. This basically tells the interpreter that
the type needs to be resolved after the class has been fully created.

This library didn't support ForwardRef checking at all. At least now, we
suport it somewhat, even if the error messages aren't perfectly useful,
we can validate that recursive and deeply nested collections are
congruent.