Create flag for auto mapping camel and snake case
JSON tends to have camel cased keys, but Python variables tend to be
snake cased. This meant that the normal use case for this library was to
have an object defined to receive the json requests and validate them,
and then you'd have to manually map everything from the camel casing, to
the exact same objects but with snake casing.

This change adds a new MappingMode keyword argument to the from/to
json/dict methods that are added by the mixin. This means that you could
have your actual dataclass defined with snake case variables, as is the
norm in Python code, but you can call
`from_dict(dct, mapping_mode=MappingMode.SnakeCase)` which will take all of
the camel case keys in your json and convert them to snake case, before
the rest of the dataclass is instantiated and type checked.