chore: Clean up CI automation
The automation itself for managing versions and how we ran flake8 and
tests was grossly out of date. Even though there's not much development
going on on this repo, I'd like to still keep it in a decent state for
anyone who wants to contribute to it in the future.